# src/main/resources/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost
    description: Generated server url
...
paths:
  "/order":
    post:
      tags:
        - Order Controller
      summary: 상품 주문
      description: 상품을 주문합니다. 결제 검증 후 주문을 진행합니다.
      operationId: order
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OrderDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/order/{orderId}/reOrder":
    post:
      tags:
        - Order Controller
      summary: 재주문
      description: 취소된 주문을 재주문합니다. 재주문 권한 검증을 거칩니다.
      operationId: reOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
  "/order/{orderId}/cancel":
    post:
      tags:
        - Order Controller
      summary: 주문 취소
      description: 특정 주문을 취소합니다. 주문 취소 권한 검증을 거칩니다.
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
  "/orders/{page}":
    get:
      tags:
        - Order Controller
      summary: 주문 내역 조회
      description: 사용자의 주문 내역을 페이지네이션으로 조회합니다.
      operationId: orderHist_1
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
    OrderDto:
      required:
        - itemId
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        count:
          maximum: 999
          minimum: 1
          type: integer
          format: int32
        totalPrice:
          type: integer
          format: int64
        makeMerchantUid:
          type: string
        impUid:
          type: string
...